[UPDATE] `token` share functions into separated files
Create:

- `create_access_token`
- `get_payload`
