+++ app/utils/__init__.py
@@ -2,6 +2,6 @@ from app.utils.get_permission_name import get_permission_name
 from app.utils.hash_password import hash_password
 from app.utils.get_response_login import get_response_login
 from app.utils.to_dict import to_dict
-from app.utils.token import create_access_token
-from app.utils.token import get_payload
+from app.utils.create_access_token import create_access_token
+from app.utils.get_payload import get_payload
 from app.utils.verify_password import verify_password
